backend/requirements.txt
asgiref==3.10.0
black==25.9.0
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
distlib==0.4.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
flake8==7.3.0
identify==2.6.15
isort==7.0.0
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
pytokens==0.1.10
PyYAML==6.0.3
redis==7.0.1
sqlparse==0.5.3
virtualenv==20.35.3
